requirements.txt
pyTelegramBotAPI>=4.32.0
python-dotenv>=1.2.2
requests>=2.32.3
Pillow>=10.0.0
